Split findMin into separate file
