+++ include/linear_function.h
@@ -1,11 +1,14 @@
 #pragma once
+#include <memory>
+#include "graph.h"
+#include "subset.h"
 
 // Defines a linear function y = ax + b
 struct LinearFunction {
   double a;
   double b;
 
-  inline double operator()(double t) {
+  inline double operator()(double t) const {
     return a * t + b;
   }
 };
@@ -15,3 +18,12 @@ struct LinearFunctionPair {
   LinearFunction first;
   LinearFunction second;
 };
+
+struct EdgeFunctions {
+  std::shared_ptr<Edge> edge;
+  LinearFunction first;
+  LinearFunction second;
+  std::shared_ptr<Subset> p1;
+  std::shared_ptr<Subset> p2;
+};
+
