+++ include/subroutine.h
@@ -23,21 +23,9 @@
 #include "graph.h"
 #include "linear_function.h"
 #include "subset.h"
+#include "find_min.h"
 
 /* ------------------------- HELPER FUNCTIONS --------------------------*/
-
-// Find min event to occur next
-// Stores minimum subset pointer in min_s and minimum edge pointer in min_e
-// Also finds event at lambda*(1+eps) (marked with _p)
-
-void findMinEvent(
-    double lambda, const std::list<std::shared_ptr<Subset>>& subsets,
-    std::list<std::shared_ptr<Edge>>& edge_functions,
-    std::unordered_map<std::shared_ptr<Edge>, LinearFunctionPair>& lin_e,
-    std::unordered_map<std::shared_ptr<Subset>, LinearFunctionPair>& lin_s,
-    std::shared_ptr<Subset>& min_s, std::shared_ptr<Edge>& min_e,
-    std::shared_ptr<Edge>& alt_e, LinearFunction& lin_val,
-    LinearFunction& lin_val_p1, LinearFunction& lin_val_p2);
 
 /* ------------------------- MAIN FUNCTIONS --------------------------*/
